+++ oxygine/src/DebugActor.cpp
@@ -154,7 +154,7 @@ namespace oxygine
 		char buff[1024] = "";
 		va_list args;
 		va_start(args, format);
-		int len = strlen(buff);
+		int len = (int)strlen(buff);
 		int i = vsnprintf(buff + len, sizeof(buff) - len, format, args);
 		va_end(args);
